+++ pymc3/theanof.py
@@ -15,9 +15,8 @@
 import numpy as np
 import theano
 
-from theano import scalar
+from theano import change_flags, scalar
 from theano import tensor as tt
-from theano.configparser import change_flags
 from theano.gof import Op
 from theano.gof.graph import inputs
 from theano.sandbox.rng_mrg import MRG_RandomStreams
@@ -440,34 +439,6 @@ def set_tt_rng(new_rng):
 
 def floatX_array(x):
     return floatX(np.array(x))
-
-
-def set_theano_conf(values):
-    """Change the theano configuration and return old values.
-
-    This is similar to `theano.configparser.change_flags`, but it
-    returns the original values in a pickleable form.
-    """
-    variables = {}
-    unknown = set(values.keys())
-    for variable in theano.configparser._config_var_list:
-        if variable.fullname in values:
-            variables[variable.fullname] = variable
-            unknown.remove(variable.fullname)
-    if len(unknown) > 0:
-        raise ValueError("Unknown theano config settings: %s" % unknown)
-
-    old = {}
-    for name, variable in variables.items():
-        old_value = variable.__get__(True, None)
-        try:
-            variable.__set__(None, values[name])
-        except Exception:
-            for key, old_value in old.items():
-                variables[key].__set__(None, old_value)
-            raise
-        old[name] = old_value
-    return old
 
 
 def ix_(*args):